Refactor frontend code for product API integration
